늑ㄷㅐ 만ㄷㅡㄴㅡㄴ중

+++ Source/PLAI/Item/Creture/Dragon/CreDraFsm.h
@@ -38,7 +38,8 @@ public:
 	void DraIdle();
 	void DraAround();
 	void DraAttack();
-	void MyTimer(void(UCreDraFsm::*Func)(), int32 time);
+	
+	void MyTimer(void(UCreDraFsm::*Func)(), float time);
 
 	void NextState();
